🚚 Rename LikeButton to WatchButton
